Move color and attribute styling out of term package.
